% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CENS_D1}
\alias{CENS_D1}
\title{Census Division 1}
\format{
A vector of FIPS codes
}
\usage{
CENS_D1
}
\description{
Census Divisions are multi-state groups, sub-sets of Census Regions.
}
\examples{
CENS_D1
}
\keyword{datasets}
