% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CENS_R1}
\alias{CENS_R1}
\title{Census Region 1}
\format{
A vector of FIPS codes
}
\usage{
CENS_R1
}
\description{
The United States is split into 4 regions: Northeast, Midwest, South and
West.
}
\examples{
CENS_R1
}
\keyword{datasets}
