% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HHS2}
\alias{HHS2}
\title{HHS Region 2}
\format{
A vector of FIPS codes
}
\usage{
HHS2
}
\description{
The Department of Health and Human Services (HHS) groups the 50 states,
the District of Columbia, and the U.S. territories into ten reporting
regions, referred to as the HHS regions.
Any number of locations can be specified here.
}
\examples{
HHS2
}
\keyword{datasets}
