% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd_provisional.R
\name{mcd_provisional}
\alias{mcd_provisional}
\title{Provisional Multiple Cause of Death, Injury Intent and Mechanism Results}
\usage{
mcd_provisional(
  wonder_url,
  save = FALSE,
  group_by_1 = MCD_GROUP_BY_1_OPTS,
  group_by_2 = MCD_GROUP_BY_2_OPTS,
  group_by_3 = MCD_GROUP_BY_2_OPTS,
  group_by_4 = MCD_GROUP_BY_2_OPTS,
  show_age_adjusted = FALSE,
  show_totals = FALSE,
  show_confidence_interval = FALSE,
  show_standard_error = FALSE,
  residence_fips = "All",
  residence_urbanization_year = c("2006", "2013"),
  residence_urbanization = URBANIZATION_OPTS,
  occurrence_fips = "All",
  occurrence_urbanization_year = c("2006", "2013"),
  occurrence_urbanization = URBANIZATION_OPTS,
  age = ALL_AGE_GROUPS_OPTS,
  gender = GENDER_OPTS,
  hispanic_origin = HISPANIC_ORIGIN_OPTS,
  race_option = RACE_OPTIONS_OPTS,
  race = "All",
  period_option = PERIOD_OPTS,
  period = "All",
  weekday = WEEKDAY_OPTS,
  autopsy = AUTOPSY_OPTS,
  place_of_death = PLACE_OF_DEATH_OPTS,
  ucd_option = MCD_UCD_OPTS,
  ucd_injury_intent = INJURY_INTENT_OPTS,
  ucd_injury_mechanism = INJURY_MECHANISM_OPTS,
  ucd_drug_alcohol = DRUG_ALCOHOL_OPTS,
  ucd_icd_codes = "All",
  ucd_cause_113 = ICD10_113_LIST_OPTS,
  mcd_option = MCD_MCD_OPTS,
  mcd_drug_alcohol = DRUG_ALCOHOL_OPTS,
  mcd_drug_alcohol_and = DRUG_ALCOHOL_OPTS,
  mcd_icd_codes = "All",
  mcd_icd_codes_and = "All",
  mcd_infant_list = "All",
  mcd_infant_list_and = "All",
  mcd_cause_113 = ICD10_113_LIST_OPTS,
  mcd_cause_113_and = ICD10_113_LIST_OPTS
)
}
\arguments{
\item{wonder_url}{CDC WONDER url with the 'session id'. See
\code{\link[=session_mcd_provisional]{session_mcd_provisional()}}}

\item{save}{Save the query instead of downloading data from WONDER}

\item{group_by_1}{One of
\itemize{
\item Residence Census Region
\item Residence Census Division
\item Residence HHS Region
\item Residence State
\item Residence County
\item Residence 2013 Urbanization
\item Residence 2006 Urbanization
\item Occurrence Census Region
\item Occurrence Census Division
\item Occurrence HHS Region
\item Occurrence State
\item Occurrence County
\item Occurrence 2013 Urbanization
\item Occurrence 2006 Urbanization
\item Ten-Year Age Groups
\item Five-Year Age Groups
\item Single-Year Ages
\item Infant Age Groups
\item Gender
\item Hispanic Origin
\item Single Race 6
\item Single Race 15
\item Single/Multi Race 31
\item Year
\item Month
\item MMWR Year
\item MMWR Week
\item Autopsy
\item Place of Death
\item UCD - 15 Leading Causes of Death
\item UCD - 15 Leading Causes of Death (Infants)
\item UCD - ICD Chapter
\item UCD - ICD Sub-Chapter
\item Underlying Cause of death
\item UCD - ICD-10 113 Cause List
\item UCD - ICD-10 130 Cause List (Infants)
\item UCD - Injury Intent
\item UCD - Injury Mechanism & All Other Leading Causes
\item UCD - Drug/Alcohol Induced
\item UCD - Drug/Alcohol Induced Cause
\item MCD - ICD Chapter
\item MCD - ICD Sub-Chapter
\item Multiple Cause of death
\item MCD - ICD-10 113 Cause List
\item MCD - ICD-10 130 Cause List (Infants)
\item MCD - Drug/Alcohol Induced
\item MCD - Drug/Alcohol Induced Cause
}}

\item{group_by_2}{One of
\itemize{
\item None
\item Residence Census Region
\item Residence Census Division
\item Residence HHS Region
\item Residence State
\item Residence County
\item Residence 2013 Urbanization
\item Residence 2006 Urbanization
\item Occurrence Census Region
\item Occurrence Census Division
\item Occurrence HHS Region
\item Occurrence State
\item Occurrence County
\item Occurrence 2013 Urbanization
\item Occurrence 2006 Urbanization
\item Ten-Year Age Groups
\item Five-Year Age Groups
\item Single-Year Ages
\item Infant Age Groups
\item Gender
\item Hispanic Origin
\item Single Race 6
\item Single Race 15
\item Single/Multi Race 31
\item Year
\item Month
\item MMWR Year
\item MMWR Week
\item Autopsy
\item Place of Death
\item UCD - 15 Leading Causes of Death
\item UCD - 15 Leading Causes of Death (Infants)
\item UCD - ICD Chapter
\item UCD - ICD Sub-Chapter
\item Underlying Cause of death
\item UCD - ICD-10 113 Cause List
\item UCD - ICD-10 130 Cause List (Infants)
\item UCD - Injury Intent
\item UCD - Injury Mechanism & All Other Leading Causes
\item UCD - Drug/Alcohol Induced
\item UCD - Drug/Alcohol Induced Cause
\item MCD - ICD Chapter
\item MCD - ICD Sub-Chapter
\item Multiple Cause of death
\item MCD - ICD-10 113 Cause List
\item MCD - ICD-10 130 Cause List (Infants)
\item MCD - Drug/Alcohol Induced
\item MCD - Drug/Alcohol Induced Cause
}}

\item{group_by_3}{One of
\itemize{
\item None
\item Residence Census Region
\item Residence Census Division
\item Residence HHS Region
\item Residence State
\item Residence County
\item Residence 2013 Urbanization
\item Residence 2006 Urbanization
\item Occurrence Census Region
\item Occurrence Census Division
\item Occurrence HHS Region
\item Occurrence State
\item Occurrence County
\item Occurrence 2013 Urbanization
\item Occurrence 2006 Urbanization
\item Ten-Year Age Groups
\item Five-Year Age Groups
\item Single-Year Ages
\item Infant Age Groups
\item Gender
\item Hispanic Origin
\item Single Race 6
\item Single Race 15
\item Single/Multi Race 31
\item Year
\item Month
\item MMWR Year
\item MMWR Week
\item Autopsy
\item Place of Death
\item UCD - 15 Leading Causes of Death
\item UCD - 15 Leading Causes of Death (Infants)
\item UCD - ICD Chapter
\item UCD - ICD Sub-Chapter
\item Underlying Cause of death
\item UCD - ICD-10 113 Cause List
\item UCD - ICD-10 130 Cause List (Infants)
\item UCD - Injury Intent
\item UCD - Injury Mechanism & All Other Leading Causes
\item UCD - Drug/Alcohol Induced
\item UCD - Drug/Alcohol Induced Cause
\item MCD - ICD Chapter
\item MCD - ICD Sub-Chapter
\item Multiple Cause of death
\item MCD - ICD-10 113 Cause List
\item MCD - ICD-10 130 Cause List (Infants)
\item MCD - Drug/Alcohol Induced
\item MCD - Drug/Alcohol Induced Cause
}}

\item{group_by_4}{One of
\itemize{
\item None
\item Residence Census Region
\item Residence Census Division
\item Residence HHS Region
\item Residence State
\item Residence County
\item Residence 2013 Urbanization
\item Residence 2006 Urbanization
\item Occurrence Census Region
\item Occurrence Census Division
\item Occurrence HHS Region
\item Occurrence State
\item Occurrence County
\item Occurrence 2013 Urbanization
\item Occurrence 2006 Urbanization
\item Ten-Year Age Groups
\item Five-Year Age Groups
\item Single-Year Ages
\item Infant Age Groups
\item Gender
\item Hispanic Origin
\item Single Race 6
\item Single Race 15
\item Single/Multi Race 31
\item Year
\item Month
\item MMWR Year
\item MMWR Week
\item Autopsy
\item Place of Death
\item UCD - 15 Leading Causes of Death
\item UCD - 15 Leading Causes of Death (Infants)
\item UCD - ICD Chapter
\item UCD - ICD Sub-Chapter
\item Underlying Cause of death
\item UCD - ICD-10 113 Cause List
\item UCD - ICD-10 130 Cause List (Infants)
\item UCD - Injury Intent
\item UCD - Injury Mechanism & All Other Leading Causes
\item UCD - Drug/Alcohol Induced
\item UCD - Drug/Alcohol Induced Cause
\item MCD - ICD Chapter
\item MCD - ICD Sub-Chapter
\item Multiple Cause of death
\item MCD - ICD-10 113 Cause List
\item MCD - ICD-10 130 Cause List (Infants)
\item MCD - Drug/Alcohol Induced
\item MCD - Drug/Alcohol Induced Cause
}}

\item{show_age_adjusted}{Show age adjusted values and confidence bands}

\item{show_totals}{Show totals}

\item{show_confidence_interval}{Show the confidence interval of the rate}

\item{show_standard_error}{Show the standard error of the rate,}

\item{residence_fips}{"All" or a list of state/county FIPS codes}

\item{residence_urbanization_year}{One of
\itemize{
\item 2006
\item 2013
}}

\item{residence_urbanization}{One or more of
\itemize{
\item All Categories
\item Large Central Metro
\item Large Fringe Metro
\item Medium Metro
\item Small Metro
\item Micropolitan (Nonmetro)
\item NonCore (Nonmetro)
}}

\item{occurrence_fips}{"All" or a list of state/county FIPS codes}

\item{occurrence_urbanization_year}{One of
\itemize{
\item 2006
\item 2013
}}

\item{occurrence_urbanization}{One or more of
\itemize{
\item All Categories
\item Large Central Metro
\item Large Fringe Metro
\item Medium Metro
\item Small Metro
\item Micropolitan (Nonmetro)
\item NonCore (Nonmetro)
}}

\item{age}{One or more of
\itemize{
\item All Ages
\item < 1
\item 1-4
\item 5-14
\item 15-24
\item 25-34
\item 35-44
\item 45-54
\item 55-64
\item 65-74
\item 75-84
\item 85+
\item Not Stated
\item 5-9
\item 10-14
\item 15-19
\item 20-24
\item 25-29
\item 30-34
\item 35-39
\item 40-44
\item 45-49
\item 50-54
\item 55-59
\item 60-64
\item 65-69
\item 70-74
\item 75-79
\item 80-84
\item 85-89
\item 90-94
\item 95-99
\item 100+
\item 0
\item 1
\item 2
\item 3
\item 4
\item 5
\item 6
\item 7
\item 8
\item 9
\item 10
\item 11
\item 12
\item 13
\item 14
\item 15
\item 16
\item 17
\item 18
\item 19
\item 20
\item 21
\item 22
\item 23
\item 24
\item 25
\item 26
\item 27
\item 28
\item 29
\item 30
\item 31
\item 32
\item 33
\item 34
\item 35
\item 36
\item 37
\item 38
\item 39
\item 40
\item 41
\item 42
\item 43
\item 44
\item 45
\item 46
\item 47
\item 48
\item 49
\item 50
\item 51
\item 52
\item 53
\item 54
\item 55
\item 56
\item 57
\item 58
\item 59
\item 60
\item 61
\item 62
\item 63
\item 64
\item 65
\item 66
\item 67
\item 68
\item 69
\item 70
\item 71
\item 72
\item 73
\item 74
\item 75
\item 76
\item 77
\item 78
\item 79
\item 80
\item 81
\item 82
\item 83
\item 84
\item 85
\item 86
\item 87
\item 88
\item 89
\item 90
\item 91
\item 92
\item 93
\item 94
\item 95
\item 96
\item 97
\item 98
\item 99
\item All Infant Ages
\item < 1 day
\item 1-6 days
\item 7-27 days
\item 28-364 days
}}

\item{gender}{One or more of
\itemize{
\item All Genders
\item Female
\item Male
}}

\item{hispanic_origin}{One or more of
\itemize{
\item All Origins
\item Not Hispanic or Latino
\item Hispanic or Latino
\item Not Stated
}}

\item{race_option}{One or more of
\itemize{
\item Single Race 6
\item Single Race 15
\item Single/Multi Race 31
}}

\item{race}{One or more of
\itemize{
\item All Races
\item American Indian or Alaska Native
\item Asian
\item Black or African American
\item Native Hawaiian or Other Pacific Islander
\item White
\item More than one race
\item Black
\item American Indian or Alaskan Native
\item Asian Indian
\item Chinese
\item Filipino
\item Japanese
\item Korean
\item Vietnamese
\item Other Asian
\item Hawaiian
\item Guamanian
\item Samoan
\item Other Pacific Islander
\item White (only)
\item Black (only)
\item AIAN (American Indian or Alaskan Native) (only)
\item Asian (only)
\item NHOPI (Native Hawaiian or Other Pacific Islander) (only)
\item Black and White
\item Black and AIAN
\item Black and Asian
\item Black and NHOPI
\item AIAN and White
\item AIAN and Asian
\item AIAN and NHOPI
\item Asian and White
\item Asian and NHOPI
\item NHOPI and White
\item Black, AIAN, and White
\item Black, AIAN, and Asian
\item Black, AIAN, and NHOPI
\item Black, Asian, and White
\item Black, Asian, and NHOPI
\item Black, NHOPI, and White
\item AIAN, Asian, and White
\item AIAN, NHOPI, and White
\item AIAN, Asian, and NHOPI
\item Asian, NHOPI, and White
\item Black, AIAN, Asian, and White
\item Black, AIAN, Asian, and NHOPI
\item Black, AIAN, NHOPI, and White
\item Black, Asian, NHOPI, and White
\item AIAN, Asian, NHOPI, and White
\item Black, AIAN, Asian, NHOPI, and White
}}

\item{period_option}{One or more of
\itemize{
\item Year
\item MMWR
}}

\item{period}{A vector of years/months/weeks in the interval
2018/01:2021/12 for months or 2018/01:2021/52 for weeks
(or last year available)}

\item{weekday}{deprecated}

\item{autopsy}{One or more of
\itemize{
\item All Values
\item No
\item Yes
\item Unknown
}}

\item{place_of_death}{One or more of
\itemize{
\item All Places
\item Medical Facility - Inpatient
\item Medical Facility - Outpatient or ER
\item Medical Facility - Dead on Arrival
\item Medical Facility - Status unknown
\item Decedent's home
\item Hospice facility
\item Nursing home/long term care
\item Other
\item Place of death unknown
}}

\item{ucd_option}{One of
\itemize{
\item ICD-10 Codes
\item ICD-10 130 Cause List (Infants)
\item Drug/Alcohol Induced Causes
\item ICD-10 113 Cause List
\item Injury Intent and Mechanism
}}

\item{ucd_injury_intent}{One or more of
\itemize{
\item All Causes of Death
\item Unintentional
\item Suicide
\item Homicide
\item Undetermined
\item Legal Intervention / Operations of War
\item Non-Injury, no intent classified
}}

\item{ucd_injury_mechanism}{One or more of
\itemize{
\item All Causes of Death
\item Cut/Pierce
\item Drowning
\item Fall
\item Fire/Flame
\item Hot object/Substance
\item Firearm
\item Machinery
\item Motor Vehicle Traffic
\item Other Pedal cyclist
\item Other Pedestrian
\item Other land transport
\item Other transport
\item Natural/Environmental
\item Overexertion
\item Poisoning
\item Struck by or against
\item Suffocation
\item Other specified, classifiable Injury
\item Other specified, not elsewhere classified Injury
\item Unspecified Injury
\item Non-Injury: Intestinal infections
\item Non-Injury: Tuberculosis
\item Non-Injury: Whooping cough
\item Non-Injury: Scarlet fever and erysipelas
\item Non-Injury: Meningococcal infection
\item Non-Injury: Septicemia
\item Non-Injury: Syphilis
\item Non-Injury: Acute poliomyelitis
\item Non-Injury: Arthropod-borne viral encephalitis
\item Non-Injury: Measles
\item Non-Injury: Viral hepatitis
\item Non-Injury: Human immunodeficiency virus (HIV) disease
\item Non-Injury: Malaria
\item Non-Injury: Other and unspecified infectious and parasitic diseases and their sequelae
\item Non-Injury: Malignant neoplasms (Cancers)
\item Non-Injury: In situ neoplasms, benign neoplasms and neoplasms of uncertain or unknown behavior
\item Non-Injury: Anemias
\item Non-Injury: Diabetes mellitus
\item Non-Injury: Nutritional Deficiencies
\item Non-Injury: Meningitis
\item Non-Injury: Parkinsons disease
\item Non-Injury: Alzheimers disease
\item Non-Injury: Diseases of Heart
\item Non-Injury: Essential (primary) hypertension and hypertensive renal disease
\item Non-Injury: Cerebrovascular diseases, including stroke
\item Non-Injury: Atherosclerosis
\item Non-Injury: Other diseases of the circulatory system
\item Non-Injury: Other disorders of circulatory system
\item Non-Injury: Influenza & Pneumonia
\item Non-Injury: Other acute lower respiratory infections
\item Non-Injury: Chronic lower respiratory diseases
\item Non-Injury: Pneumoconioses and chemical effects
\item Non-Injury: Pneumonitis due to solids and liquids
\item Non-Injury: Other diseases of respiratory system
\item Non-Injury: Peptic ulcer
\item Non-Injury: Diseases of appendix
\item Non-Injury: Hernia
\item Non-Injury: Chronic liver disease and cirrhosis
\item Non-Injury: Cholelithiasis and other disorders of gallbladder
\item Non-Injury: Nephritis, nephrotic syndrome and nephrosis
\item Non-Injury: Infections of kidney
\item Non-Injury: Hyperplasia of prostate
\item Non-Injury: Inflammatory diseases of female pelvic organs
\item Non-Injury: Pregnancy with abortive outcome
\item Non-Injury: Other complications of pregnancy, childbirth and the puerperium
\item Non-Injury: Certain conditions originating in the perinatal period
\item Non-Injury: Congenital malformations, deformations and chromosomal abnormalities
\item Non-Injury: Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified
\item Non-Injury: All other diseases (Residual)
\item Non-Injury: Complications of medical and surgical care
}}

\item{ucd_drug_alcohol}{One or more of
\itemize{
\item All Causes of Death
\item Drug-induced causes
\item Drug poisonings (overdose) Unintentional (X40-X44)
\item Drug poisonings (overdose) Suicide (X60-X64)
\item Drug poisonings (overdose) Homicide (X85)
\item Drug poisonings (overdose) Undetermined (Y10-Y14)
\item All other drug-induced causes
\item Alcohol-induced causes
\item Alcohol poisonings (overdose) (X45, X65, Y15)
\item All other alcohol-induced causes
\item O
\item All other non-drug and non-alcohol causes
}}

\item{ucd_icd_codes}{An ICD-10 code:
Visit \url{https://wonder.cdc.gov/mcd-icd10-provisional.html}
for the complete list}

\item{ucd_cause_113}{One or more of
\itemize{
\item All Causes of Death
\item #Salmonella infections (A01-A02)
\item #Shigellosis and amebiasis (A03,A06)
\item Certain other intestinal infections (A04,A07-A09)
\item #Tuberculosis (A16-A19)
\item Respiratory tuberculosis (A16)
\item Other tuberculosis (A17-A19)
\item #Whooping cough (A37)
\item #Scarlet fever and erysipelas (A38,A46)
\item #Meningococcal infection (A39)
\item #Septicemia (A40-A41)
\item #Syphilis (A50-A53)
\item #Acute poliomyelitis (A80)
\item #Arthropod-borne viral encephalitis (A83-A84,A85.2)
\item #Measles (B05)
\item #Viral hepatitis (B15-B19)
\item #Human immunodeficiency virus (HIV) disease (B20-B24)
\item #Malaria (B50-B54)
\item Other and unspecified infectious and parasitic diseases and their sequelae (A00,A05,A20-A36,A42-A44,A48-A49,A54-A79,A81-A82,A85.0-A85.1,A85.8,A86-B04,B06-B09,B25-B49,B55-B99,U07.1)
\item #Malignant neoplasms (C00-C97)
\item Malignant neoplasms of lip, oral cavity and pharynx (C00-C14)
\item Malignant neoplasm of esophagus (C15)
\item Malignant neoplasm of stomach (C16)
\item Malignant neoplasms of colon, rectum and anus (C18-C21)
\item Malignant neoplasms of liver and intrahepatic bile ducts (C22)
\item Malignant neoplasm of pancreas (C25)
\item Malignant neoplasm of larynx (C32)
\item Malignant neoplasms of trachea, bronchus and lung (C33-C34)
\item Malignant melanoma of skin (C43)
\item Malignant neoplasm of breast (C50)
\item Malignant neoplasm of cervix uteri (C53)
\item Malignant neoplasms of corpus uteri and uterus, part unspecified (C54-C55)
\item Malignant neoplasm of ovary (C56)
\item Malignant neoplasm of prostate (C61)
\item Malignant neoplasms of kidney and renal pelvis (C64-C65)
\item Malignant neoplasm of bladder (C67)
\item Malignant neoplasms of meninges, brain and other parts of central nervous system (C70-C72)
\item Malignant neoplasms of lymphoid, hematopoietic and related tissue (C81-C96)
\item Hodgkin disease (C81)
\item Non-Hodgkin lymphoma (C82-C85)
\item Leukemia (C91-C95)
\item Multiple myeloma and immunoproliferative neoplasms (C88,C90)
\item Other and unspecified malignant neoplasms of lymphoid, hematopoietic and related tissue (C96)
\item All other and unspecified malignant neoplasms (C17,C23-C24,C26-C31,C37-C41,C44-C49,C51-C52,C57-C60,C62-C63,C66,C68-C69,C73-C80,C97)
\item #In situ neoplasms, benign neoplasms and neoplasms of uncertain or unknown behavior (D00-D48)
\item #Anemias (D50-D64)
\item #Diabetes mellitus (E10-E14)
\item #Nutritional deficiencies (E40-E64)
\item Malnutrition (E40-E46)
\item Other nutritional deficiencies (E50-E64)
\item #Meningitis (G00,G03)
\item #Parkinson disease (G20-G21)
\item #Alzheimer disease (G30)
\item Major cardiovascular diseases (I00-I78)
\item #Diseases of heart (I00-I09,I11,I13,I20-I51)
\item Acute rheumatic fever and chronic rheumatic heart diseases (I00-I09)
\item Hypertensive heart disease (I11)
\item Hypertensive heart and renal disease (I13)
\item Ischemic heart diseases (I20-I25)
\item Acute myocardial infarction (I21-I22)
\item Other acute ischemic heart diseases (I24)
\item Other forms of chronic ischemic heart disease (I20,I25)
\item Atherosclerotic cardiovascular disease, so described (I25.0)
\item All other forms of chronic ischemic heart disease (I20,I25.1-I25.9)
\item Other heart diseases (I26-I51)
\item Acute and subacute endocarditis (I33)
\item Diseases of pericardium and acute myocarditis (I30-I31,I40)
\item Heart failure (I50)
\item All other forms of heart disease (I26-I28,I34-I38,I42-I49,I51)
\item #Essential hypertension and hypertensive renal disease (I10,I12,I15)
\item #Cerebrovascular diseases (I60-I69)
\item #Atherosclerosis (I70)
\item Other diseases of circulatory system (I71-I78)
\item #Aortic aneurysm and dissection (I71)
\item Other diseases of arteries, arterioles and capillaries (I72-I78)
\item Other disorders of circulatory system (I80-I99)
\item #Influenza and pneumonia (J09-J18)
\item Influenza (J09-J11)
\item Pneumonia (J12-J18)
\item Other acute lower respiratory infections (J20-J22,U04)
\item #Acute bronchitis and bronchiolitis (J20-J21)
\item Other and unspecified acute lower respiratory infections (J22,U04)
\item #Chronic lower respiratory diseases (J40-J47)
\item Bronchitis, chronic and unspecified (J40-J42)
\item Emphysema (J43)
\item Asthma (J45-J46)
\item Other chronic lower respiratory diseases (J44,J47)
\item #Pneumoconioses and chemical effects (J60-J66,J68,U07.0)
\item #Pneumonitis due to solids and liquids (J69)
\item Other diseases of respiratory system (J00-J06,J30- J39,J67,J70-J98)
\item #Peptic ulcer (K25-K28)
\item #Diseases of appendix (K35-K38)
\item #Hernia (K40-K46)
\item #Chronic liver disease and cirrhosis (K70,K73-K74)
\item Alcoholic liver disease (K70)
\item Other chronic liver disease and cirrhosis (K73-K74)
\item #Cholelithiasis and other disorders of gallbladder (K80-K82)
\item #Nephritis, nephrotic syndrome and nephrosis (N00-N07,N17-N19,N25-N27)
\item Acute and rapidly progressive nephritic and nephrotic syndrome (N00-N01,N04)
\item Chronic glomerulonephritis, nephritis and nephropathy not specified as acute or chronic, and renal sclerosis unspecified (N02-N03,N05-N07,N26)
\item Renal failure (N17-N19)
\item Other disorders of kidney (N25,N27)
\item #Infections of kidney (N10-N12,N13.6,N15.1)
\item #Hyperplasia of prostate (N40)
\item #Inflammatory diseases of female pelvic organs (N70-N76)
\item #Pregnancy, childbirth and the puerperium (O00-O99)
\item Pregnancy with abortive outcome (O00-O07)
\item Other complications of pregnancy, childbirth and the puerperium (O10-O99)
\item #Certain conditions originating in the perinatal period (P00-P96)
\item #Congenital malformations, deformations and chromosomal abnormalities (Q00-Q99)
\item Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified (R00-R99)
\item All other diseases (Residual)
\item #Accidents (unintentional injuries) (V01-X59,Y85-Y86)
\item Transport accidents (V01-V99,Y85)
\item Motor vehicle accidents (V02-V04,V09.0,V09.2,V12-V14,V19.0-V19.2,V19.4-V19.6,V20-V79,V80.3-V80.5,V81.0-V81.1,V82.0-V82.1,V83-V86,V87.0-V87.8,V88.0-V88.8,V89.0,V89.2)
\item Other land transport accidents (V01,V05-V06,V09.1,V09.3-V09.9,V10-V11,V15-V18,V19.3,V19.8-V19.9,V80.0-V80.2,V80.6-V80.9,V81.2-V81.9,V82.2-V82.9,V87.9,V88.9,V89.1,V89.3,V89.9)
\item Water, air and space, and other and unspecified transport accidents and their sequelae (V90-V99,Y85)
\item Nontransport accidents (W00-X59,Y86)
\item Falls (W00-W19)
\item Accidental discharge of firearms (W32-W34)
\item Accidental drowning and submersion (W65-W74)
\item Accidental exposure to smoke, fire and flames (X00-X09)
\item Accidental poisoning and exposure to noxious substances (X40-X49)
\item Other and unspecified nontransport accidents and their sequelae (W20-W31,W35-W64,W75-W99,X10-X39,X50-X59,Y86)
\item #Intentional self-harm (suicide) (*U03,X60-X84,Y87.0)
\item Intentional self-harm (suicide) by discharge of firearms (X72-X74)
\item Intentional self-harm (suicide) by other and unspecified means and their sequelae (*U03,X60-X71,X75-X84,Y87.0)
\item #Assault (homicide) (*U01-*U02,X85-Y09,Y87.1)
\item Assault (homicide) by discharge of firearms (*U01.4,X93-X95)
\item Assault (homicide) by other and unspecified means and their sequelae (*U01.0-*U01.3,*U01.5-*U01.9,*U02,X85-X92,X96-Y09,Y87.1)
\item #Legal intervention (Y35,Y89.0)
\item Events of undetermined intent (Y10-Y34,Y87.2,Y89.9)
\item Discharge of firearms, undetermined intent (Y22-Y24)
\item Other and unspecified events of undetermined intent and their sequelae (Y10-Y21,Y25-Y34,Y87.2,Y89.9)
\item #Operations of war and their sequelae (Y36,Y89.1)
\item #Complications of medical and surgical care (Y40-Y84,Y88)
\item #Enterocolitis due to Clostridium difficile (A04.7)
\item #COVID-19 (U07.1)
}}

\item{mcd_option}{One or more of
\itemize{
\item MCD - ICD-10 Codes
\item MCD - ICD-10 130 Cause List (Infants)
\item MCD - Drug/Alcohol Induced Causes
\item MCD - ICD-10 113 Cause List
}}

\item{mcd_drug_alcohol}{One or more of
\itemize{
\item All Causes of Death
\item Drug-induced causes
\item Drug poisonings (overdose) Unintentional (X40-X44)
\item Drug poisonings (overdose) Suicide (X60-X64)
\item Drug poisonings (overdose) Homicide (X85)
\item Drug poisonings (overdose) Undetermined (Y10-Y14)
\item All other drug-induced causes
\item Alcohol-induced causes
\item Alcohol poisonings (overdose) (X45, X65, Y15)
\item All other alcohol-induced causes
\item O
\item All other non-drug and non-alcohol causes
}}

\item{mcd_drug_alcohol_and}{One or more of
\itemize{
\item All Causes of Death
\item Drug-induced causes
\item Drug poisonings (overdose) Unintentional (X40-X44)
\item Drug poisonings (overdose) Suicide (X60-X64)
\item Drug poisonings (overdose) Homicide (X85)
\item Drug poisonings (overdose) Undetermined (Y10-Y14)
\item All other drug-induced causes
\item Alcohol-induced causes
\item Alcohol poisonings (overdose) (X45, X65, Y15)
\item All other alcohol-induced causes
\item O
\item All other non-drug and non-alcohol causes
}}

\item{mcd_icd_codes}{An ICD-10 code: See
\url{https://wonder.cdc.gov/mcd-icd10-provisional.html} for the complete list}

\item{mcd_icd_codes_and}{An ICD-10 code: See
\url{https://wonder.cdc.gov/mcd-icd10-provisional.html} for the complete
list}

\item{mcd_infant_list}{Not implemented}

\item{mcd_infant_list_and}{Not implemented}

\item{mcd_cause_113}{One or more of
\itemize{
\item All Causes of Death
\item #Salmonella infections (A01-A02)
\item #Shigellosis and amebiasis (A03,A06)
\item Certain other intestinal infections (A04,A07-A09)
\item #Tuberculosis (A16-A19)
\item Respiratory tuberculosis (A16)
\item Other tuberculosis (A17-A19)
\item #Whooping cough (A37)
\item #Scarlet fever and erysipelas (A38,A46)
\item #Meningococcal infection (A39)
\item #Septicemia (A40-A41)
\item #Syphilis (A50-A53)
\item #Acute poliomyelitis (A80)
\item #Arthropod-borne viral encephalitis (A83-A84,A85.2)
\item #Measles (B05)
\item #Viral hepatitis (B15-B19)
\item #Human immunodeficiency virus (HIV) disease (B20-B24)
\item #Malaria (B50-B54)
\item Other and unspecified infectious and parasitic diseases and their sequelae (A00,A05,A20-A36,A42-A44,A48-A49,A54-A79,A81-A82,A85.0-A85.1,A85.8,A86-B04,B06-B09,B25-B49,B55-B99,U07.1)
\item #Malignant neoplasms (C00-C97)
\item Malignant neoplasms of lip, oral cavity and pharynx (C00-C14)
\item Malignant neoplasm of esophagus (C15)
\item Malignant neoplasm of stomach (C16)
\item Malignant neoplasms of colon, rectum and anus (C18-C21)
\item Malignant neoplasms of liver and intrahepatic bile ducts (C22)
\item Malignant neoplasm of pancreas (C25)
\item Malignant neoplasm of larynx (C32)
\item Malignant neoplasms of trachea, bronchus and lung (C33-C34)
\item Malignant melanoma of skin (C43)
\item Malignant neoplasm of breast (C50)
\item Malignant neoplasm of cervix uteri (C53)
\item Malignant neoplasms of corpus uteri and uterus, part unspecified (C54-C55)
\item Malignant neoplasm of ovary (C56)
\item Malignant neoplasm of prostate (C61)
\item Malignant neoplasms of kidney and renal pelvis (C64-C65)
\item Malignant neoplasm of bladder (C67)
\item Malignant neoplasms of meninges, brain and other parts of central nervous system (C70-C72)
\item Malignant neoplasms of lymphoid, hematopoietic and related tissue (C81-C96)
\item Hodgkin disease (C81)
\item Non-Hodgkin lymphoma (C82-C85)
\item Leukemia (C91-C95)
\item Multiple myeloma and immunoproliferative neoplasms (C88,C90)
\item Other and unspecified malignant neoplasms of lymphoid, hematopoietic and related tissue (C96)
\item All other and unspecified malignant neoplasms (C17,C23-C24,C26-C31,C37-C41,C44-C49,C51-C52,C57-C60,C62-C63,C66,C68-C69,C73-C80,C97)
\item #In situ neoplasms, benign neoplasms and neoplasms of uncertain or unknown behavior (D00-D48)
\item #Anemias (D50-D64)
\item #Diabetes mellitus (E10-E14)
\item #Nutritional deficiencies (E40-E64)
\item Malnutrition (E40-E46)
\item Other nutritional deficiencies (E50-E64)
\item #Meningitis (G00,G03)
\item #Parkinson disease (G20-G21)
\item #Alzheimer disease (G30)
\item Major cardiovascular diseases (I00-I78)
\item #Diseases of heart (I00-I09,I11,I13,I20-I51)
\item Acute rheumatic fever and chronic rheumatic heart diseases (I00-I09)
\item Hypertensive heart disease (I11)
\item Hypertensive heart and renal disease (I13)
\item Ischemic heart diseases (I20-I25)
\item Acute myocardial infarction (I21-I22)
\item Other acute ischemic heart diseases (I24)
\item Other forms of chronic ischemic heart disease (I20,I25)
\item Atherosclerotic cardiovascular disease, so described (I25.0)
\item All other forms of chronic ischemic heart disease (I20,I25.1-I25.9)
\item Other heart diseases (I26-I51)
\item Acute and subacute endocarditis (I33)
\item Diseases of pericardium and acute myocarditis (I30-I31,I40)
\item Heart failure (I50)
\item All other forms of heart disease (I26-I28,I34-I38,I42-I49,I51)
\item #Essential hypertension and hypertensive renal disease (I10,I12,I15)
\item #Cerebrovascular diseases (I60-I69)
\item #Atherosclerosis (I70)
\item Other diseases of circulatory system (I71-I78)
\item #Aortic aneurysm and dissection (I71)
\item Other diseases of arteries, arterioles and capillaries (I72-I78)
\item Other disorders of circulatory system (I80-I99)
\item #Influenza and pneumonia (J09-J18)
\item Influenza (J09-J11)
\item Pneumonia (J12-J18)
\item Other acute lower respiratory infections (J20-J22,U04)
\item #Acute bronchitis and bronchiolitis (J20-J21)
\item Other and unspecified acute lower respiratory infections (J22,U04)
\item #Chronic lower respiratory diseases (J40-J47)
\item Bronchitis, chronic and unspecified (J40-J42)
\item Emphysema (J43)
\item Asthma (J45-J46)
\item Other chronic lower respiratory diseases (J44,J47)
\item #Pneumoconioses and chemical effects (J60-J66,J68,U07.0)
\item #Pneumonitis due to solids and liquids (J69)
\item Other diseases of respiratory system (J00-J06,J30- J39,J67,J70-J98)
\item #Peptic ulcer (K25-K28)
\item #Diseases of appendix (K35-K38)
\item #Hernia (K40-K46)
\item #Chronic liver disease and cirrhosis (K70,K73-K74)
\item Alcoholic liver disease (K70)
\item Other chronic liver disease and cirrhosis (K73-K74)
\item #Cholelithiasis and other disorders of gallbladder (K80-K82)
\item #Nephritis, nephrotic syndrome and nephrosis (N00-N07,N17-N19,N25-N27)
\item Acute and rapidly progressive nephritic and nephrotic syndrome (N00-N01,N04)
\item Chronic glomerulonephritis, nephritis and nephropathy not specified as acute or chronic, and renal sclerosis unspecified (N02-N03,N05-N07,N26)
\item Renal failure (N17-N19)
\item Other disorders of kidney (N25,N27)
\item #Infections of kidney (N10-N12,N13.6,N15.1)
\item #Hyperplasia of prostate (N40)
\item #Inflammatory diseases of female pelvic organs (N70-N76)
\item #Pregnancy, childbirth and the puerperium (O00-O99)
\item Pregnancy with abortive outcome (O00-O07)
\item Other complications of pregnancy, childbirth and the puerperium (O10-O99)
\item #Certain conditions originating in the perinatal period (P00-P96)
\item #Congenital malformations, deformations and chromosomal abnormalities (Q00-Q99)
\item Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified (R00-R99)
\item All other diseases (Residual)
\item #Accidents (unintentional injuries) (V01-X59,Y85-Y86)
\item Transport accidents (V01-V99,Y85)
\item Motor vehicle accidents (V02-V04,V09.0,V09.2,V12-V14,V19.0-V19.2,V19.4-V19.6,V20-V79,V80.3-V80.5,V81.0-V81.1,V82.0-V82.1,V83-V86,V87.0-V87.8,V88.0-V88.8,V89.0,V89.2)
\item Other land transport accidents (V01,V05-V06,V09.1,V09.3-V09.9,V10-V11,V15-V18,V19.3,V19.8-V19.9,V80.0-V80.2,V80.6-V80.9,V81.2-V81.9,V82.2-V82.9,V87.9,V88.9,V89.1,V89.3,V89.9)
\item Water, air and space, and other and unspecified transport accidents and their sequelae (V90-V99,Y85)
\item Nontransport accidents (W00-X59,Y86)
\item Falls (W00-W19)
\item Accidental discharge of firearms (W32-W34)
\item Accidental drowning and submersion (W65-W74)
\item Accidental exposure to smoke, fire and flames (X00-X09)
\item Accidental poisoning and exposure to noxious substances (X40-X49)
\item Other and unspecified nontransport accidents and their sequelae (W20-W31,W35-W64,W75-W99,X10-X39,X50-X59,Y86)
\item #Intentional self-harm (suicide) (*U03,X60-X84,Y87.0)
\item Intentional self-harm (suicide) by discharge of firearms (X72-X74)
\item Intentional self-harm (suicide) by other and unspecified means and their sequelae (*U03,X60-X71,X75-X84,Y87.0)
\item #Assault (homicide) (*U01-*U02,X85-Y09,Y87.1)
\item Assault (homicide) by discharge of firearms (*U01.4,X93-X95)
\item Assault (homicide) by other and unspecified means and their sequelae (*U01.0-*U01.3,*U01.5-*U01.9,*U02,X85-X92,X96-Y09,Y87.1)
\item #Legal intervention (Y35,Y89.0)
\item Events of undetermined intent (Y10-Y34,Y87.2,Y89.9)
\item Discharge of firearms, undetermined intent (Y22-Y24)
\item Other and unspecified events of undetermined intent and their sequelae (Y10-Y21,Y25-Y34,Y87.2,Y89.9)
\item #Operations of war and their sequelae (Y36,Y89.1)
\item #Complications of medical and surgical care (Y40-Y84,Y88)
\item #Enterocolitis due to Clostridium difficile (A04.7)
\item #COVID-19 (U07.1)
}}

\item{mcd_cause_113_and}{One or more of
\itemize{
\item All Causes of Death
\item #Salmonella infections (A01-A02)
\item #Shigellosis and amebiasis (A03,A06)
\item Certain other intestinal infections (A04,A07-A09)
\item #Tuberculosis (A16-A19)
\item Respiratory tuberculosis (A16)
\item Other tuberculosis (A17-A19)
\item #Whooping cough (A37)
\item #Scarlet fever and erysipelas (A38,A46)
\item #Meningococcal infection (A39)
\item #Septicemia (A40-A41)
\item #Syphilis (A50-A53)
\item #Acute poliomyelitis (A80)
\item #Arthropod-borne viral encephalitis (A83-A84,A85.2)
\item #Measles (B05)
\item #Viral hepatitis (B15-B19)
\item #Human immunodeficiency virus (HIV) disease (B20-B24)
\item #Malaria (B50-B54)
\item Other and unspecified infectious and parasitic diseases and their sequelae (A00,A05,A20-A36,A42-A44,A48-A49,A54-A79,A81-A82,A85.0-A85.1,A85.8,A86-B04,B06-B09,B25-B49,B55-B99,U07.1)
\item #Malignant neoplasms (C00-C97)
\item Malignant neoplasms of lip, oral cavity and pharynx (C00-C14)
\item Malignant neoplasm of esophagus (C15)
\item Malignant neoplasm of stomach (C16)
\item Malignant neoplasms of colon, rectum and anus (C18-C21)
\item Malignant neoplasms of liver and intrahepatic bile ducts (C22)
\item Malignant neoplasm of pancreas (C25)
\item Malignant neoplasm of larynx (C32)
\item Malignant neoplasms of trachea, bronchus and lung (C33-C34)
\item Malignant melanoma of skin (C43)
\item Malignant neoplasm of breast (C50)
\item Malignant neoplasm of cervix uteri (C53)
\item Malignant neoplasms of corpus uteri and uterus, part unspecified (C54-C55)
\item Malignant neoplasm of ovary (C56)
\item Malignant neoplasm of prostate (C61)
\item Malignant neoplasms of kidney and renal pelvis (C64-C65)
\item Malignant neoplasm of bladder (C67)
\item Malignant neoplasms of meninges, brain and other parts of central nervous system (C70-C72)
\item Malignant neoplasms of lymphoid, hematopoietic and related tissue (C81-C96)
\item Hodgkin disease (C81)
\item Non-Hodgkin lymphoma (C82-C85)
\item Leukemia (C91-C95)
\item Multiple myeloma and immunoproliferative neoplasms (C88,C90)
\item Other and unspecified malignant neoplasms of lymphoid, hematopoietic and related tissue (C96)
\item All other and unspecified malignant neoplasms (C17,C23-C24,C26-C31,C37-C41,C44-C49,C51-C52,C57-C60,C62-C63,C66,C68-C69,C73-C80,C97)
\item #In situ neoplasms, benign neoplasms and neoplasms of uncertain or unknown behavior (D00-D48)
\item #Anemias (D50-D64)
\item #Diabetes mellitus (E10-E14)
\item #Nutritional deficiencies (E40-E64)
\item Malnutrition (E40-E46)
\item Other nutritional deficiencies (E50-E64)
\item #Meningitis (G00,G03)
\item #Parkinson disease (G20-G21)
\item #Alzheimer disease (G30)
\item Major cardiovascular diseases (I00-I78)
\item #Diseases of heart (I00-I09,I11,I13,I20-I51)
\item Acute rheumatic fever and chronic rheumatic heart diseases (I00-I09)
\item Hypertensive heart disease (I11)
\item Hypertensive heart and renal disease (I13)
\item Ischemic heart diseases (I20-I25)
\item Acute myocardial infarction (I21-I22)
\item Other acute ischemic heart diseases (I24)
\item Other forms of chronic ischemic heart disease (I20,I25)
\item Atherosclerotic cardiovascular disease, so described (I25.0)
\item All other forms of chronic ischemic heart disease (I20,I25.1-I25.9)
\item Other heart diseases (I26-I51)
\item Acute and subacute endocarditis (I33)
\item Diseases of pericardium and acute myocarditis (I30-I31,I40)
\item Heart failure (I50)
\item All other forms of heart disease (I26-I28,I34-I38,I42-I49,I51)
\item #Essential hypertension and hypertensive renal disease (I10,I12,I15)
\item #Cerebrovascular diseases (I60-I69)
\item #Atherosclerosis (I70)
\item Other diseases of circulatory system (I71-I78)
\item #Aortic aneurysm and dissection (I71)
\item Other diseases of arteries, arterioles and capillaries (I72-I78)
\item Other disorders of circulatory system (I80-I99)
\item #Influenza and pneumonia (J09-J18)
\item Influenza (J09-J11)
\item Pneumonia (J12-J18)
\item Other acute lower respiratory infections (J20-J22,U04)
\item #Acute bronchitis and bronchiolitis (J20-J21)
\item Other and unspecified acute lower respiratory infections (J22,U04)
\item #Chronic lower respiratory diseases (J40-J47)
\item Bronchitis, chronic and unspecified (J40-J42)
\item Emphysema (J43)
\item Asthma (J45-J46)
\item Other chronic lower respiratory diseases (J44,J47)
\item #Pneumoconioses and chemical effects (J60-J66,J68,U07.0)
\item #Pneumonitis due to solids and liquids (J69)
\item Other diseases of respiratory system (J00-J06,J30- J39,J67,J70-J98)
\item #Peptic ulcer (K25-K28)
\item #Diseases of appendix (K35-K38)
\item #Hernia (K40-K46)
\item #Chronic liver disease and cirrhosis (K70,K73-K74)
\item Alcoholic liver disease (K70)
\item Other chronic liver disease and cirrhosis (K73-K74)
\item #Cholelithiasis and other disorders of gallbladder (K80-K82)
\item #Nephritis, nephrotic syndrome and nephrosis (N00-N07,N17-N19,N25-N27)
\item Acute and rapidly progressive nephritic and nephrotic syndrome (N00-N01,N04)
\item Chronic glomerulonephritis, nephritis and nephropathy not specified as acute or chronic, and renal sclerosis unspecified (N02-N03,N05-N07,N26)
\item Renal failure (N17-N19)
\item Other disorders of kidney (N25,N27)
\item #Infections of kidney (N10-N12,N13.6,N15.1)
\item #Hyperplasia of prostate (N40)
\item #Inflammatory diseases of female pelvic organs (N70-N76)
\item #Pregnancy, childbirth and the puerperium (O00-O99)
\item Pregnancy with abortive outcome (O00-O07)
\item Other complications of pregnancy, childbirth and the puerperium (O10-O99)
\item #Certain conditions originating in the perinatal period (P00-P96)
\item #Congenital malformations, deformations and chromosomal abnormalities (Q00-Q99)
\item Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified (R00-R99)
\item All other diseases (Residual)
\item #Accidents (unintentional injuries) (V01-X59,Y85-Y86)
\item Transport accidents (V01-V99,Y85)
\item Motor vehicle accidents (V02-V04,V09.0,V09.2,V12-V14,V19.0-V19.2,V19.4-V19.6,V20-V79,V80.3-V80.5,V81.0-V81.1,V82.0-V82.1,V83-V86,V87.0-V87.8,V88.0-V88.8,V89.0,V89.2)
\item Other land transport accidents (V01,V05-V06,V09.1,V09.3-V09.9,V10-V11,V15-V18,V19.3,V19.8-V19.9,V80.0-V80.2,V80.6-V80.9,V81.2-V81.9,V82.2-V82.9,V87.9,V88.9,V89.1,V89.3,V89.9)
\item Water, air and space, and other and unspecified transport accidents and their sequelae (V90-V99,Y85)
\item Nontransport accidents (W00-X59,Y86)
\item Falls (W00-W19)
\item Accidental discharge of firearms (W32-W34)
\item Accidental drowning and submersion (W65-W74)
\item Accidental exposure to smoke, fire and flames (X00-X09)
\item Accidental poisoning and exposure to noxious substances (X40-X49)
\item Other and unspecified nontransport accidents and their sequelae (W20-W31,W35-W64,W75-W99,X10-X39,X50-X59,Y86)
\item #Intentional self-harm (suicide) (*U03,X60-X84,Y87.0)
\item Intentional self-harm (suicide) by discharge of firearms (X72-X74)
\item Intentional self-harm (suicide) by other and unspecified means and their sequelae (*U03,X60-X71,X75-X84,Y87.0)
\item #Assault (homicide) (*U01-*U02,X85-Y09,Y87.1)
\item Assault (homicide) by discharge of firearms (*U01.4,X93-X95)
\item Assault (homicide) by other and unspecified means and their sequelae (*U01.0-*U01.3,*U01.5-*U01.9,*U02,X85-X92,X96-Y09,Y87.1)
\item #Legal intervention (Y35,Y89.0)
\item Events of undetermined intent (Y10-Y34,Y87.2,Y89.9)
\item Discharge of firearms, undetermined intent (Y22-Y24)
\item Other and unspecified events of undetermined intent and their sequelae (Y10-Y21,Y25-Y34,Y87.2,Y89.9)
\item #Operations of war and their sequelae (Y36,Y89.1)
\item #Complications of medical and surgical care (Y40-Y84,Y88)
\item #Enterocolitis due to Clostridium difficile (A04.7)
\item #COVID-19 (U07.1)
}}
}
\value{
A data.frame with the data from wonder
}
\description{
Download Provisional Mortality Statistics, 2018 through
Last Month death certificate data from
\url{https://wonder.cdc.gov/mcd-icd10-provisional.html)}
}
