% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder_session.R
\name{session_mcd_final18}
\alias{session_mcd_final18}
\title{Agree to the CDC WONDER data use restrictions}
\usage{
session_mcd_final18(I_Agree = FALSE)
}
\arguments{
\item{I_Agree}{Pre-agree to the WONDER data use restrictions}
}
\value{
An URL string including the session id
}
\description{
Agree to the CDC WONDER data use restrictions
}
