% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_STATES_FIPS}
\alias{US_STATES_FIPS}
\title{Dataset with FIPS codes for US states}
\format{
An object of class \code{data.frame} with 51 rows and 3 columns.
}
\usage{
US_STATES_FIPS
}
\description{
\itemize{
\item \code{state_name}: Title-case name of state
\item \code{state_fips}: State FIPS code (2-digit, 0-padded, character)
\item \code{state_usps}: Upper-case UNITED STATES POSTAL SERVICE
abbreviations for states
}
}
\keyword{datasets}
